refactor : divide Controller
